backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
black==25.9.0
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
cryptography==46.0.3
distro==1.9.0
ecdsa==0.19.1
fastapi==0.119.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.0
mypy_extensions==1.1.0
openai==2.4.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytokens==0.2.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
wheel==0.45.1
pdfplumber
reportlab
PyMuPDF>=1.24.0
pytesseract>=0.3.10
Pillow>=10.0.0

# Added for RegWatch, Evidence Bundle, Rule Library
pyyaml>=6.0
jsonschema>=4.20.0
requests>=2.31.0
beautifulsoup4>=4.12.0
apscheduler>=3.10.0
feedparser>=6.0.0

# Added for GİB Real Data Services (KURGAN)
aiohttp>=3.9.0
openpyxl>=3.1.0
lxml>=5.0.0

# AI providers
anthropic>=0.40.0
tenacity>=8.2.0
